Moved mimeType back to all mutation sources
